refactor(tests): split large test files by tested concept
